Renamed app name from "myapplication" to "QArmy"
