+++ app/build.gradle
@@ -3,11 +3,11 @@ plugins {
 }
 
 android {
-    namespace 'com.example.myapplication'
+    namespace 'com.example.QArmy'
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.example.myapplication"
+        applicationId "com.example.QArmy"
         minSdk 24
         targetSdk 33
         versionCode 1
