+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MyApplication"
+        android:theme="@style/Theme.QArmy"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
