+++ app/src/main/java/com/example/QArmy/Image.java
@@ -1,0 +1,4 @@
+package com.example.QArmy;
+
+public class Image {
+}
