+++ app/src/main/java/com/example/QArmy/PlayerProfile.java
@@ -1,0 +1,4 @@
+package com.example.QArmy;
+
+public class PlayerProfile {
+}
